packages/xsft-corridor/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xsft-corridor"
version = "0.1.0"
description = "Leak-aware export corridor with budget and invariance checks, producing signed receipts."
authors = [{name = "Steven Lanier"}, {name="SEAL Division"}]
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.9"
dependencies = ["numpy>=1.24","pydantic>=2.6","pyyaml>=6.0"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]

[project.scripts]
xsft = "xsft_corridor.cli:main"
